[cross-table][fw] Handle the cross-table position update and rollback.
A generic system has been implemented: when a sdcard data is loaded or
when a "holes on a line" generator is used, a vector of PulseUpdate is
created.

The main position menu handles the vector of PulseUpdate to give the
possibility to the user to move the table or rollback it at each
step (ie for each PulseUpdate in the vector).